Feature: query task events in task

+++ orm/event.go
@@ -24,7 +24,7 @@ func NewEventOrm(db *gorm.DB) *EventOrm {
 	}
 }
 
-func (e *EventOrm) GetEventsByTaskID(taskID int) ([]*Event, error) {
+func (e *EventOrm) GetEventsByTaskID(taskID int64) ([]*Event, error) {
 	var events []*Event
 	if err := e.DB.Model(Event{}).Where("task_id = ?", taskID).Find(&events).Error; err != nil {
 		return nil, err
